Move the effect class into the PostProcessor header
